requirement.txt
bcrypt==3.1.4
Django==1.11.7
djangorestframework==3.7.3
uuid==1.30